Added a resource folder that'll hold files I'll be reading and writing to. More of why is explained in other commits. But in short, I'm creating objects through the usage of text files holding values that'll represent the attributes and later on build the object itself. I'm hopefully as well going to try out on of the VG criterias that involves being able to save a gaming session to file, to be able to pick it up later on and continue game.
AddAnimal and AddFood are text files meant to be used during runtime, when youre running a game and you want to add more objects.

Animals and Food are text files holding the values representing the objects that'll be created when the program is run initially.

The text files holding animals go by the following format AND IS VERY IMPORTANT TO FOLLOW IN ORDER FOR THE PROGRAM TO RUN PROPERLY:
[CLASS/ANIMALGROUP][SPECIE][AMOUNT:CREATE][WEIGHTRANGE][AGE][PRICE]

The text files holding foods go by the following format AND IS VERY IMPORTANT TO FOLLOW IN ORDER FOR THE PROGRAM TO RUN PROPERLY
[CLASS/FOODGROUP][FOOD:NAME][WEIGHT][PRICE/KG]