Animal class is coming to its finish, it's not completely finished as it needs a couple of methods more I assume which I haven't written yet, and some need finishing.
As of now, the method for healing the animal when it eats doesn't work as I want yet. As of now it works as the following: for each kg of food the animal get a 10 percent health increase. I want to change this to have a more realistic effect to it, as a mouse doesn't need as much of food as an elephant needs.